Fix finding the CSS with Express static hosting. Note the paths are relative. It's somewhat strange to look at when starting out, but take a look at how I changed the pathnames to make this work.
While here, add a simple .gitignore that ensures that node_modules/
doesn't get committed (saving you lots of time and disk space!) as well
as files that end in ~ since those are backup files left behind by Unix
text editors (Vim, in my case).

Remove the node_modules/ directory now that it will be ignored. That's
what the package.json file is for!